requirements.txt
annotated-types==0.7.0
anyio==4.11.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
fastapi==0.117.1
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.1
packageurl-python==0.17.5
packaging==25.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.4.0
pluggy==1.6.0
py-serializable==2.1.0
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
toml==0.10.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
